Fix session completion detection
The key is to know when a session is complete.
This is tricky because there may be missing data at then end.
The only way we can use to check if there is no more data for a session
is when there is data available after a session end date.

Use this assumption to fix session completion on the watch.